Problem 35. (#20)
Added `sieve_of_eratosthenes` function.

Earlier, I had removed this function and added an iterator which iterates over prime numbers. However, for P35, it is useful to have the actual sieve, because there will be many primality checks on numbers over a wide range of values. Hence, I moved the part of the code which generates the sieve into a separate function. I also optimised the prime iterator by directly returning a generator rather than collecting the primes into a vector and then returning an iterator over it. And noted the suboptimality of using the sieve of Eratosthenes and the primes iterator together.