Massive optimisation in P23.
For some reason not quite clear to me, the approach I though was optimal (binary-searching for the number in a sorted vector which, when added to a particular number, yields an abundant sum) turned out to not be optimal. (Perhaps because of repetition?) Instead, adding all possible pairs to find all abundant sums, and then complementing that set is much faster! The running time went down from 480 ms to 60 ms. Wow.