Problem 45 (#46)
Added an inverter to the polygonal number iterator and used it to solve P45 and also updated the solution of P44 to use the same. Interestingly, the running time of the latter decreased from 30 ms to 22 ms for some reason, although the check for pentagonal numbers is the same. The decrease could be the result of generating pentagonal numbers using additions instead of multiplications.