P12 optimisation (#51)
Found this trick in the thread for P12. Reduces running time from 100 ms to 20 ms. Another optimisation removes an extra calculation: divisors of one factor are already calculated in the previous loop iteration. Using it instead of recalculating reduces the time to 10 ms.

Counting the divisors using `Divisors` is faster than using `PrimeDivisors`, but I will keep it because it may be useful for some other problem later on.