Faster solution by exploiting triplet generator.
The Pythagorean triplet generating algorithm can be adapted to generate all triplets with a sum less than or equal to a given number instead of those with a sum equal to the same (which is what `utils::PythagoreanTriplets` does). Consequently, running the adapted algorithm for 1500000 is faster than running the latter algorithm for all numbers from 12 to 1500000.