Made `Polygonal::invert` return `Result` instead of `Option`.
Even if the index is not an integer, it may still help find out what the nearest index might be. Actually, I made this change to try another approach in P85 (keep increasing `a` and determine what value `b` must take so that the number of rectangles is 2000000), but ended up not using it because I am too lazy to figure out how to make it work.